Split into subdir, fix mnemonic to ed25519
