+++ nym-vpn-core/crates/nym-vpn-api-client/src/types/mod.rs
@@ -1,0 +1,15 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+mod account;
+mod device;
+mod gateway;
+
+#[cfg(test)]
+mod test_fixtures;
+
+pub use account::VpnApiAccount;
+pub use device::Device;
+pub use gateway::{GatewayMinPerformance, GatewayType};
+
+pub use nym_contracts_common::Percent;
