+++ nym-vpn-core/crates/nym-vpn-api-client/src/types/test_fixtures.rs
@@ -1,0 +1,12 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+// The default acccount mnemonic, the same as in the js integration tests
+pub(super) const DEFAULT_MNEMONIC: &str =
+       "range mystery picture decline olympic acoustic lesson quick rebuild panda royal fold start leader egg hammer width olympic worry length crawl couch link mobile";
+pub(super) const DEFAULT_MNEMONIC_ID: &str = "n1sslaag27wfydyrvyua72hg5e0vteglxrs8nw3c";
+
+// The default device mnemonic, the same as in the js integration tests
+pub(super) const DEFAULT_DEVICE_MNEMONIC: &str =
+        "pitch deputy proof fire movie put bread ribbon what chef zebra car vacuum gadget steak board state oyster layer glory barely thrive nice box";
+pub(super) const DEFAULT_DEVICE_IDENTITY_KEY: &str = "FJDUECYAeosXhNGjxf8w5MJM7N2DfDwQznvWwTxJz6ft";
